Move up SetBreakpoint command
